+++ e2e-tests/tests/remix.test.ts
@@ -1,4 +1,3 @@
-/* eslint-disable jest/expect-expect */
 import * as path from 'node:path';
 import { Integration } from '../../lib/Constants';
 import {
@@ -17,6 +16,7 @@ import {
   revertLocalChanges,
   startWizardInstance,
 } from '../utils';
+import { afterAll, beforeAll, describe, test } from 'vitest';
 
 const SERVER_TEMPLATE = `import { createRequestHandler } from '@remix-run/express';
 import { installGlobals } from '@remix-run/node';